+++ pkg/agents/team1/agent2/agent.go
@@ -3,21 +3,21 @@ package agent
 import (
 	"log"
 
-	. "github.com/SOMAS2021/SOMAS2021/pkg/agents/default"
+	"github.com/SOMAS2021/SOMAS2021/pkg/agents"
 )
 
 type CustomAgent2 struct {
-	*BaseAgent
+	*agents.Base
 	myNumber int
 }
 
-func New(baseAgent *BaseAgent) (Agent, error) {
+func New(baseAgent *agents.Base) (agents.Agent, error) {
 	return &CustomAgent2{
-		BaseAgent: baseAgent,
-		myNumber:  0,
+		Base:     baseAgent,
+		myNumber: 0,
 	}, nil
 }
 
 func (a *CustomAgent2) Run() {
-	log.Printf("Custom agent team 2 has floor: %d", a.GetFloor())
+	log.Printf("Custom agent team 2 has floor: %d", a.Floor())
 }
